blog_Project/requirement.txt
alembic==1.4.0
aliyun-python-sdk-core==2.13.15
blinker==1.4
blueprint==3.4.2
bp==0.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
importlib-metadata==1.5.0
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
Mako==1.1.1
MarkupSafe==1.1.1
Pillow==7.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
zipp==3.0.0
